# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactKind": {
        "description": "Роль файла внутри распознанного пакета.",
        "enum": [
          "blocks_json",
          "results_md",
          "results_html",
          "package_zip",
          "other"
        ],
        "title": "ArtifactKind",
        "type": "string"
      },
...
      "ContentUrl": {
        "description": "Ссылка на файл ревизии прямо из хранилища.\n\nAPI не проксирует бинарные данные: документ на 50–500 МБ не должен проходить через\nпроцесс приложения (ADR-0002).",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "expires_in": {
            "exclusiveMinimum": 0.0,
            "title": "Expires In",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_in",
          "filename",
          "content_type"
        ],
        "title": "ContentUrl",
        "type": "object"
      },
      "DocumentKind": {
        "description": "Природа документа. BIM-форматы принимаются на хранение, но не разбираются.",
        "enum": [
          "pdf",
          "recognized_package",
          "revit",
          "navisworks",
          "ifc",
          "other"
        ],
        "title": "DocumentKind",
        "type": "string"
      },
      "DocumentRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discipline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discipline"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "document_kind": {
            "$ref": "#/components/schemas/DocumentKind"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "display_name",
          "discipline",
          "document_kind",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentRead",
        "type": "object"
      },
      "DocumentRevisionRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "processing_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error Code"
          },
          "processing_status": {
            "$ref": "#/components/schemas/ProcessingStatus"
          },
          "revision_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Label"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "source_metadata": {
            "additionalProperties": true,
            "title": "Source Metadata",
            "type": "object"
          },
          "source_mime": {
            "title": "Source Mime",
            "type": "string"
          },
          "source_sha256": {
            "title": "Source Sha256",
            "type": "string"
          },
          "source_size": {
            "title": "Source Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "document_id",
          "revision_label",
          "source_filename",
          "source_mime",
          "source_size",
          "source_sha256",
          "processing_status",
          "processing_error_code",
          "source_metadata",
          "created_at"
        ],
        "title": "DocumentRevisionRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_DocumentRevisionRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentRevisionRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[DocumentRevisionRead]",
        "type": "object"
      },
...
      "Page_RegionRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegionRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RegionRead]",
        "type": "object"
      },
      "Page_SheetRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SheetRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[SheetRead]",
        "type": "object"
      },
      "ProcessingStatus": {
        "description": "Состояние обработки конкретной ревизии.\n\nЕдинственное изменяемое поле ревизии: описывает не файл, а ход работы над ним.",
        "enum": [
          "pending",
          "unprocessed",
          "processor_unavailable",
          "importing",
          "ready",
          "failed"
        ],
        "title": "ProcessingStatus",
        "type": "string"
      },
...
      "RecognitionArtifactRead": {
        "properties": {
          "artifact_kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          }
        },
        "required": [
          "id",
          "revision_id",
          "artifact_kind",
          "schema_version",
          "sha256",
          "source_filename",
          "metadata",
          "created_at"
        ],
        "title": "RecognitionArtifactRead",
        "type": "object"
      },
      "RegionRead": {
        "description": "Область распознавания.\n\ncoords_norm — [x0, y0, x1, y1] от левого верхнего угла листа, значения в [0, 1].\nПеревод в экранные координаты выполняет клиент при отрисовке (ADR-0008).",
        "properties": {
          "block_type": {
            "title": "Block Type",
            "type": "string"
          },
          "coords_norm": {
            "items": {
              "type": "number"
            },
            "title": "Coords Norm",
            "type": "array"
          },
          "external_block_id": {
            "title": "External Block Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "legacy_metadata": {
            "additionalProperties": true,
            "title": "Legacy Metadata",
            "type": "object"
          },
          "ordinal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordinal"
          },
          "polygon_points": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygon Points"
          },
          "raw_content_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Content Md"
          },
          "recognition_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recognition Status"
          },
          "shape_type": {
            "$ref": "#/components/schemas/RegionShape"
          },
          "sheet_id": {
            "format": "uuid",
            "title": "Sheet Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sheet_id",
          "external_block_id",
          "ordinal",
          "block_type",
          "shape_type",
          "coords_norm",
          "polygon_points",
          "recognition_status",
          "raw_content_md",
          "legacy_metadata"
        ],
        "title": "RegionRead",
        "type": "object"
      },
      "RegionShape": {
        "enum": [
          "rectangle",
          "polygon"
        ],
        "title": "RegionShape",
        "type": "string"
      },
...
      "SheetRead": {
        "properties": {
          "height_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Px"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page_index": {
            "title": "Page Index",
            "type": "integer"
          },
          "page_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Label"
          },
          "region_count": {
            "minimum": 0.0,
            "title": "Region Count",
            "type": "integer"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "rotation": {
            "title": "Rotation",
            "type": "integer"
          },
          "width_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width Px"
          }
        },
        "required": [
          "id",
          "revision_id",
          "page_index",
          "page_label",
          "width_px",
          "height_px",
          "rotation",
          "region_count"
        ],
        "title": "SheetRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API портала автоматизированного подсчёта строительных объёмов. Stage 1: проекты, импорт распознанных пакетов и просмотрщик документации. Расчёта объёмов и вызовов моделей на этом этапе нет.",
    "title": "Quantor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{document_id}": {
      "get": {
        "operationId": "read_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Документ",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/revisions": {
      "get": {
        "operationId": "list_document_revisions",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "description": "Сколько записей вернуть",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Сколько записей вернуть",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Сколько записей пропустить",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Сколько записей пропустить",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DocumentRevisionRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ревизии документа",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/revisions/{revision_id}": {
      "get": {
        "operationId": "read_revision",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRevisionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ревизия документа",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/revisions/{revision_id}/artifacts": {
      "get": {
        "operationId": "list_revision_artifacts",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecognitionArtifactRead"
                  },
                  "title": "Response List Revision Artifacts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Артефакты распознавания",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/revisions/{revision_id}/content-url": {
      "get": {
        "description": "Временная ссылка прямо в хранилище.\n\nПросмотрщику нужен доступ с поддержкой Range-запросов, чтобы читать только текущую\nстраницу документа. Проксировать такой файл через процесс приложения нельзя.",
        "operationId": "read_revision_content_url",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ссылка на файл ревизии",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/revisions/{revision_id}/sheets": {
      "get": {
        "operationId": "list_revision_sheets",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "description": "Сколько записей вернуть",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Сколько записей вернуть",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Сколько записей пропустить",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Сколько записей пропустить",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SheetRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Листы ревизии",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/sheets/{sheet_id}/regions": {
      "get": {
        "operationId": "list_sheet_regions",
        "parameters": [
          {
            "in": "path",
            "name": "sheet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sheet Id",
              "type": "string"
            }
          },
          {
            "description": "Сколько записей вернуть",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Сколько записей вернуть",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Сколько записей пропустить",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Сколько записей пропустить",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Фильтр по типу области",
            "in": "query",
            "name": "block_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по типу области",
              "title": "Block Type"
            }
          },
          {
            "description": "Фильтр по статусу распознавания",
            "in": "query",
            "name": "recognition_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по статусу распознавания",
              "title": "Recognition Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RegionRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Области на листе",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}